pyproject.toml
[project]
name = "media-gathering"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "shift" }
]
dependencies = [
    "beautifulsoup4>=4.12.3",
    "coverage>=7.5.3",
    "cryptography>=42.0.7",
    "emoji>=2.12.1",
    "freezegun>=1.5.1",
    "httpx>=0.27.0",
    "isort>=5.13.2",
    "jinja2>=3.1.4",
    "mock>=5.1.0",
    "orjson>=3.10.3",
    "pillow>=10.3.0",
    "pixivpy3>=3.7.5",
    "plyer>=2.1.0",
    "pycodestyle>=2.11.1",
    "ruff>=0.4.6",
    "slack-sdk>=3.27.2",
    "sqlalchemy>=2.0.30",
    "twitter-api-client>=0.10.22",
    "xmltodict>=0.13.0",
]
readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.isort]
line_length = 119
multi_line_output = 9

[tool.ruff]
exclude = [
    "__pycache__",
    ".coveragerc",
    ".git",
    ".venv",
    "venv",
]
line-length = 119

[tool.ruff.lint]
fixable = ["ALL"]
select = ["E", "F"]
ignore = ["F401", "F541", "F841"]
preview = true

[tool.ruff.format]
preview = true
quote-style = "double"

[tool.rye]
managed = true
dev-dependencies = []

[tool.rye.scripts]
unittest = "python -m unittest"
coverage_html = {chain = ["coverage run --source . -m unittest discover", "coverage html"]}

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/media_gathering"]
